refactor : PieceGroup으로 메서드 및 초기화 이동
- Board에서 매개인자로 받아 PieceGroup을 초기화 하던 방법을
PieceGroup에서 모든 기물을 초기화하는 방법으로 변경하였습니다.

- 기존에 모든 기물을 초기화하던 initialize 메서드를 제거하였습니다.
