+++ src/main/java/cooper/chess/Board.java
@@ -11,65 +11,11 @@ public class Board {
     private final PieceGroup pieceGroup;
 
     public Board() {
-        this.pieceGroup = new PieceGroup(createPawnGroup());
-    }
-
-    private Map<Color, List<Piece>> createPawnGroup() {
-        Map<Color, List<Piece>> pawnListMap = new HashMap<>();
-
-        pawnListMap.put(Color.WHITE, new ArrayList<>());
-        pawnListMap.put(Color.BLACK, new ArrayList<>());
-
-        return pawnListMap;
-    }
-
-    public void initialize() {
-        initWhitePiece();
-        initBlackPiece();
-    }
-
-    private void initWhitePiece() {
-        List<Piece> pieceList = pieceGroup.getPawnList(Color.WHITE);
-
-        pieceList.add(Piece.createWhiteRook());
-        pieceList.add(Piece.createWhiteKnight());
-        pieceList.add(Piece.createWhiteBishop());
-        pieceList.add(Piece.createWhiteQueen());
-        pieceList.add(Piece.createWhiteKing());
-        pieceList.add(Piece.createWhiteBishop());
-        pieceList.add(Piece.createWhiteKnight());
-        pieceList.add(Piece.createWhiteRook());
-
-        for (int pawnCount = 0; pawnCount < BOARD_SIZE; pawnCount++) {
-            pieceList.add(Piece.createWhitePawn());
-        }
-    }
-
-    private void initBlackPiece() {
-        List<Piece> pieceList = pieceGroup.getPawnList(Color.BLACK);
-
-        for (int pawnCount = 0; pawnCount < BOARD_SIZE; pawnCount++) {
-            pieceList.add(Piece.createBlackPawn());
-        }
-
-        pieceList.add(Piece.createBlackRook());
-        pieceList.add(Piece.createBlackKnight());
-        pieceList.add(Piece.createBlackBishop());
-        pieceList.add(Piece.createBlackQueen());
-        pieceList.add(Piece.createBlackKing());
-        pieceList.add(Piece.createBlackBishop());
-        pieceList.add(Piece.createBlackKnight());
-        pieceList.add(Piece.createBlackRook());
+        this.pieceGroup = new PieceGroup();
     }
 
     public void add(Piece piece) {
-        List<Piece> pieceList = pieceGroup.getPawnList(piece.getColor());
-
-        if (pieceList.size() >= 8) {
-            return;
-        }
-
-        pieceList.add(piece);
+        pieceGroup.add(piece);
     }
 
     public int pieceCount() {
@@ -77,17 +23,7 @@ public class Board {
     }
 
     public Piece findPawn(int index, Color color) {
-        List<Piece> pieceList = pieceGroup.getPawnList(color);
-
-        if (pieceList.size() == 0) {
-            throw new IllegalArgumentException("size가 0입니다.");
-        }
-
-        if ((0 > index) || (index >= pieceList.size())) {
-            throw new IllegalArgumentException("범위를 벗어나는 인덱스 입니다.");
-        }
-
-        return pieceList.get(index);
+        return pieceGroup.findPawn(index, color);
     }
 
     public String showBoard() {
