+++ src/main/java/cooper/chess/ChessGame.java
@@ -9,10 +9,6 @@ public class ChessGame {
     private final Scanner scanner = new Scanner(System.in);
     private final Board board = new Board();
 
-    public void init() {
-        board.initialize();
-    }
-
     public void play() {
         boolean isPlaying = true;
