+++ src/main/java/cooper/chess/Main.java
@@ -3,8 +3,6 @@ package cooper.chess;
 public class Main {
     public static void main(String[] args) {
         ChessGame chessGame = new ChessGame();
-
-        chessGame.init();
         chessGame.play();
     }
 }
